Modifying ignore system in link.sh

+++ link.sh
@@ -1,37 +1,60 @@
 #!/usr/bin/env bash
 
-# Script configs
-IGNORE="bashrc|bash_profile|zshrc|ssh|link|gitmodules|pathogen_submodule"
+# This array will contain any strings for directories you want to ignore.
+declare -a IGNORE
+
+# Obviously we don't want to link this script.
+IGNORE+=('link.sh')
 
 # Get current directory
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
+# This directory will contain any custom extension scripts.
 EXT_DIR="$DIR/.link_ext"
 
+# Now we run any custom extensions.
 for i in $(ls $EXT_DIR)
 do
 	source $EXT_DIR/$i
 done
 
+pushd $DIR &> /dev/null
+
+# to the ignore array.
 echo -e "\e[1;35mSymlinking all other config files:\e[m"
-cd $DIR
 for file in $(git ls-files | egrep -v $IGNORE)
 do
-	if [ "$(readlink ~/.$file)" != "$DIR/$file" ]
+	ignoreThis=false
+
+	for word in ${IGNORE[@]}
+	do
+		if echo $file | grep -q $word
+		then
+			ignoreThis=true
+			break
+		fi
+	done
+
+	if ! [ ignoreThis ]
 	then
-		echo $file
-		if test ! -d `dirname ~/.$file`
+		if [ "$(readlink ~/.$file)" != "$DIR/$file" ]
 		then
-			mkdir -p `dirname ~/.$file`
+			echo $file
+			if test ! -d `dirname ~/.$file`
+			then
+				mkdir -p `dirname ~/.$file`
+			fi
+			if test -h ~/.$file
+			then
+				unlink ~/.$file
+			fi
+			rm -rf ~/.file 2>&1 >/dev/null
+			ln -sf $DIR/$file ~/.$file
 		fi
-		if test -h ~/.$file
-		then
-			unlink ~/.$file
-		fi
-		rm -rf ~/.file 2>&1 >/dev/null
-		ln -sf $DIR/$file ~/.$file
 	fi
 done
 
+# Last, we install a post-merge hook to keep everything up to date.
 if ! [ -f .git/hooks/post-merge ]
 then
 	echo "Installing post merge hook"
